black lint

+++ src/db/database_psql.py
@@ -1,7 +1,8 @@
 import sys
-from os import environ  
+from os import environ
 from collections import Counter
 import psycopg2
+
 
 class DatabasePsql:
     def __init__(self):
@@ -32,7 +33,7 @@ class DatabasePsql:
                 user=db_user,
                 password=db_pass,
                 host=db_host,
-                port=db_port    
+                port=db_port,
             )
         except psycopg2.Error as e:
             print("DB error: " + str(e))
@@ -40,10 +41,8 @@ class DatabasePsql:
 
         return conn
 
-
     def get_counters(self):
         return self.counters
-
 
     def in_blacklist(self, user_id):
         conn = self.open_connection()
@@ -52,9 +51,7 @@ class DatabasePsql:
 
         cursor = conn.cursor()
 
-        sql =   "SELECT 1 " \
-                "FROM {} " \
-                "WHERE user_id = {};"
+        sql = "SELECT 1 " "FROM {} " "WHERE user_id = {};"
 
         cursor.execute(sql.format(self.table_blacklist, user_id))
 
@@ -64,7 +61,6 @@ class DatabasePsql:
 
         return blacklisted
 
-
     def update_name(self, id, name):
         conn = self.open_connection()
         if conn == False:
@@ -72,16 +68,17 @@ class DatabasePsql:
 
         cursor = conn.cursor()
 
-        sql =   "INSERT INTO {0} (id, name) " \
-                "VALUES ({1}, '{2}') " \
-                "ON CONFLICT (id) DO UPDATE " \
-                "SET name = '{2}';"        
+        sql = (
+            "INSERT INTO {0} (id, name) "
+            "VALUES ({1}, '{2}') "
+            "ON CONFLICT (id) DO UPDATE "
+            "SET name = '{2}';"
+        )
 
         cursor.execute(sql.format(self.table_name, id, name))
 
         print(self.table_name, id, name)
         conn.commit()
-
 
     def add_blacklist(self, user_id):
         conn = self.open_connection()
@@ -90,21 +87,29 @@ class DatabasePsql:
 
         cursor = conn.cursor()
 
-        sql =   "INSERT INTO {0} " \
-                "VALUES ({4});" \
-                " " \
-                "DELETE FROM {1} " \
-                "WHERE id={4}; " \
-                "DELETE FROM {2} " \
-                "WHERE user_id={4}; " \
-                "DELETE FROM {3} " \
-                "WHERE user_id={4}; " \
-        
-        cursor.execute(sql.format(self.table_blacklist, self.table_name, self.table_counter, self.table_word, user_id))
+        sql = (
+            "INSERT INTO {0} "
+            "VALUES ({4});"
+            " "
+            "DELETE FROM {1} "
+            "WHERE id={4}; "
+            "DELETE FROM {2} "
+            "WHERE user_id={4}; "
+            "DELETE FROM {3} "
+            "WHERE user_id={4}; "
+        )
+        cursor.execute(
+            sql.format(
+                self.table_blacklist,
+                self.table_name,
+                self.table_counter,
+                self.table_word,
+                user_id,
+            )
+        )
 
         conn.commit()
         conn.close()
-
 
     def remove_blacklist(self, user_id):
         conn = self.open_connection()
@@ -113,13 +118,11 @@ class DatabasePsql:
 
         cursor = conn.cursor()
 
-        sql =   "DELETE from {} " \
-                "WHERE user_id = {};"
+        sql = "DELETE from {} " "WHERE user_id = {};"
 
         cursor.execute(sql.format(self.table_blacklist, user_id))
         conn.commit()
         conn.close()
-
 
     # inkrementoidaan. Jossei riviä ole, lisätään se
     def increment_counter(self, user_id, chat_id, counter, amount):
@@ -129,15 +132,18 @@ class DatabasePsql:
 
         cursor = conn.cursor()
 
-        sql =   "INSERT INTO {0} (user_id, chat_id, {1}) " \
-                "VALUES ({2}, {3}, {4}) " \
-                "ON CONFLICT (user_id, chat_id) DO UPDATE " \
-                "SET {1} = {0}.{1} + {4};"
+        sql = (
+            "INSERT INTO {0} (user_id, chat_id, {1}) "
+            "VALUES ({2}, {3}, {4}) "
+            "ON CONFLICT (user_id, chat_id) DO UPDATE "
+            "SET {1} = {0}.{1} + {4};"
+        )
 
-        cursor.execute(sql.format(self.table_counter, counter, user_id, chat_id, amount))
+        cursor.execute(
+            sql.format(self.table_counter, counter, user_id, chat_id, amount)
+        )
         conn.commit()
         conn.close()
-
 
     # palauttaa käyttäjä, chätti parin laskurin
     def get_counter_user(self, user_id, chat_id, counter):
@@ -147,18 +153,15 @@ class DatabasePsql:
 
         cursor = conn.cursor()
 
-        sql =   "SELECT {} " \
-                "FROM {} " \
-                "WHERE user_id = {} AND chat_id = {};"
+        sql = "SELECT {} " "FROM {} " "WHERE user_id = {} AND chat_id = {};"
 
         cursor.execute(sql.format(counter, self.table_counter, user_id, chat_id))
 
         pari = cursor.fetchone()[0]
 
         conn.close()
-        
-        return pari
 
+        return pari
 
     # nimitaulusta nimet, countterista laskurin arvo
     # kursori antaa ne tupleina -> muutetaan dictiin ja palautetaan
@@ -169,16 +172,22 @@ class DatabasePsql:
 
         cursor = conn.cursor()
 
-        sql =   "SELECT {0}.{2}, {1}.name " \
-                "FROM {0} " \
-                "INNER JOIN {1} " \
-                "ON {0}.user_id = {1}.id " \
-                "WHERE {0}.chat_id={3} " \
-                "AND {0}.{2}!=0 " \
-                "ORDER BY {0}.{2} DESC " \
-                "LIMIT {4};"
+        sql = (
+            "SELECT {0}.{2}, {1}.name "
+            "FROM {0} "
+            "INNER JOIN {1} "
+            "ON {0}.user_id = {1}.id "
+            "WHERE {0}.chat_id={3} "
+            "AND {0}.{2}!=0 "
+            "ORDER BY {0}.{2} DESC "
+            "LIMIT {4};"
+        )
 
-        cursor.execute(sql.format(self.table_counter, self.table_name, counter, chat_id, top_amount))
+        cursor.execute(
+            sql.format(
+                self.table_counter, self.table_name, counter, chat_id, top_amount
+            )
+        )
 
         res = cursor.fetchall()
 
@@ -186,20 +195,21 @@ class DatabasePsql:
 
         return dict((y, x) for x, y in res)
 
-        
     # user, chat, sana yhdistelmät uniikkeja
     # jos ei löydy -> lisätään, jos löytyy, lisätään amount counttiin
     def word_collection_add(self, user_id, chat_id, word, amount):
         conn = self.open_connection()
         if conn == False:
             return
-            
+
         cursor = conn.cursor()
 
-        sql =   "INSERT INTO {0} (user_id, chat_id, word, count) " \
-                "VALUES ({1}, {2}, '{3}', {4}) " \
-                "ON CONFLICT (user_id, chat_id, word) DO UPDATE " \
-                "SET count = {0}.count + {4}; "
+        sql = (
+            "INSERT INTO {0} (user_id, chat_id, word, count) "
+            "VALUES ({1}, {2}, '{3}', {4}) "
+            "ON CONFLICT (user_id, chat_id, word) DO UPDATE "
+            "SET count = {0}.count + {4}; "
+        )
 
         cursor.execute(sql.format(self.table_word, user_id, chat_id, word, amount))
         conn.commit()
